pyproject.toml
[tool.poetry]
name = "cptool"
version = "0.1.0"
description = ""
authors = ["Tho Nguyen"]

[tool.poetry.dependencies]
python = "^3.6"
beautifulsoup4 = "^4.9.0"
"ruamel.yaml" = "^0.16.10"

[tool.poetry.dev-dependencies]
flake8 = "*"
flake8-docstrings = "*"
flake8-rst-docstrings = "*"
mypy = "^0.770"
isort = "^4.3.21"
black = "==19.10b0"
pre-commit = "*"
ipython = "^7.13.0"
pytest = "^5.4.1"
pytest-cov = "^2.8.1"
pytest-mock = "^3.1.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
